Add edit mode stage 2
When the user clicks on a blank square it changes to a black piece.
When the user clicks on a black piece it changes to white.
When the user clicks on a white piece it changes to blank.

[next prompt]
When I remove editMode from the URL the board goes back to an initialBoard. I want it to stay the same as what I edited.